+++ src/index.js
@@ -6,25 +6,19 @@ require('string.prototype.startswith');
 //      .config('http://993be79068aa427295767e9bcda03c1c@sentry.317hu.com/15')
 //      .install();
 //      
-window[`url$Raven`] = `http://993be79068aa427295767e9bcda03c1c@sentry.317hu.com/15`;
+// window[`url$Raven`] = `http://993be79068aa427295767e9bcda03c1c@sentry.317hu.com/15`;
 require('@317hu/GlobalRavenCaptureException');
-
-import React, { Component } from 'react';
-import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-// import PropTypes from 'prop-types';
-// import createClass from 'create-react-class';
-import { configureStore, history } from './store/configureStore';
-import { saveCache } from './core/_utils/storage';
-import App from '../examples/v2.2.0/views/app';
-// import App from './views/app';
-
-// import AnimationExample from './views/AnimationExample';
-
 require('./assets/css/lib.css');
 require('./assets/css/app.css');
 // require('../examples/v2.1.1/views/app-antd.js'); // 以允许应用不同版本的 antd 样式（引用以入口为准，属于 lib-app.css）
 // require('./views/app-antd.js');
+
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import { configureStore, history } from './store/configureStore';
+import { saveCache } from './core/_utils/storage';
+import App from '../examples/v2.2.1/views/app'; // import App from './views/app';
 
 if ('scrollRestoration' in history) {
   // history.scrollRestoration = 'manual';
@@ -54,7 +48,6 @@ class Root extends Component {
     }
   }
   render () {
-    // var props = this.props;
     return (
       <Provider store={this.state.store}>
 		    <App history={history} />